Added Scoring and Background
...but it won't run the game now because it cannot find the GUIText.
Followed the instructions on the website that said:
"Unity made major changes to their GUI (Graphical User Interface) with
release Unity 4.6. One unexpected change was a redesign of the Create
menu that now makes it more challenging to create the old GUI elements.
To fix this, anywhere that the text asks you to choose GameObject >
Create Other > GUIText from the menu bar, instead:

1. From the menu bar, choose Game Object > Create Empty.
2. With that empty GameObject selected, from the menu bar, choose
Component > Rendering > GUIText.
This should set you up properly for the GUIText information in the
book."
and followed everything in the book word for word up to p.615 but again,
an issue with Unity upgrading its whole UI system...